Benchmarking contended locks only, with and without dropping the GIL. [skip ci]
Possible Conclusion: Only drop the GIL if you'll be doing a lot of work. Droping and acquiring it again is very expensive
and gets more expensive the more contested it is. Python 3 is much worse than Python 2 at this.

I say possible because it could potentially still be the lock; I need to run a test of a C++ function that does minimal work to narrow down exactly which it is.